refactor: 장르 추천 도메인 서비스 분리
