+++ backend/src/main/java/com/digginroom/digginroom/domain/recommend/GenreRecommender.java
@@ -1,0 +1,8 @@
+package com.digginroom.digginroom.domain.recommend;
+
+import com.digginroom.digginroom.domain.Genre;
+
+public interface GenreRecommender {
+
+    Genre recommend(final Long memberId);
+}
