+++ backend/src/main/java/com/digginroom/digginroom/domain/recommend/RoomRecommender.java
@@ -3,43 +3,23 @@ package com.digginroom.digginroom.domain.recommend;
 import com.digginroom.digginroom.domain.Genre;
 import com.digginroom.digginroom.domain.room.Room;
 import com.digginroom.digginroom.exception.RecommendException.NoRecommendableRoomException;
-import com.digginroom.digginroom.exception.RecommendException.UnderBoundWeightException;
-import com.digginroom.digginroom.membergenre.domain.MemberGenre;
 import com.digginroom.digginroom.repository.RoomRepository;
-import java.util.List;
-import java.util.concurrent.ThreadLocalRandom;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+
+import java.util.List;
+import java.util.concurrent.ThreadLocalRandom;
 
 @Component
 @RequiredArgsConstructor
 public class RoomRecommender {
 
     private final RoomRepository roomRepository;
+    private final GenreRecommender genreRecommender;
 
-    public Room recommend(final List<MemberGenre> memberGenres) {
-        Genre recommenedGenre = recommendGenre(memberGenres);
+    public Room recommend(final Long memberId) {
+        Genre recommenedGenre = genreRecommender.recommend(memberId);
         return recommendRoom(recommenedGenre);
-    }
-
-    private Genre recommendGenre(final List<MemberGenre> memberGenres) {
-        int weightSum = memberGenres.stream().mapToInt(MemberGenre::getWeight).sum();
-        if (weightSum <= 0) {
-            throw new UnderBoundWeightException(weightSum);
-        }
-        return pickByWeight(memberGenres, weightSum);
-    }
-
-    private Genre pickByWeight(final List<MemberGenre> memberGenres, final int weightSum) {
-        int randomizedFactor = ThreadLocalRandom.current().nextInt(weightSum);
-        int cursor = 0;
-        for (MemberGenre weightedGenre : memberGenres) {
-            cursor += weightedGenre.getWeight();
-            if (randomizedFactor < cursor) {
-                return weightedGenre.getGenre();
-            }
-        }
-        return memberGenres.get(memberGenres.size() - 1).getGenre();
     }
 
     private Room recommendRoom(final Genre recommendedGenre) {
