+++ backend/src/main/java/com/digginroom/digginroom/membergenre/domain/MemberGenreRepository.java
@@ -1,9 +1,14 @@
 package com.digginroom.digginroom.membergenre.domain;
 
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface MemberGenreRepository extends JpaRepository<MemberGenre, Long> {
+
+    default MemberGenres getAllByMemberId(final Long memberId) {
+        return new MemberGenres(findByMemberId(memberId));
+    }
 
     List<MemberGenre> findByMemberId(final Long memberId);
