+++ backend/src/main/java/com/digginroom/digginroom/service/RoomService.java
@@ -3,9 +3,7 @@ package com.digginroom.digginroom.service;
 import com.digginroom.digginroom.domain.member.Member;
 import com.digginroom.digginroom.domain.recommend.RoomRecommender;
 import com.digginroom.digginroom.domain.room.Room;
-import com.digginroom.digginroom.exception.MemberException.EmptyFavoriteException;
 import com.digginroom.digginroom.exception.RecommendException;
-import com.digginroom.digginroom.membergenre.domain.MemberGenre;
 import com.digginroom.digginroom.membergenre.domain.MemberGenreRepository;
 import com.digginroom.digginroom.membergenre.domain.vo.WeightFactor;
 import com.digginroom.digginroom.repository.MemberRepository;
@@ -13,7 +11,6 @@ import com.digginroom.digginroom.repository.RoomRepository;
 import com.digginroom.digginroom.service.dto.RoomResponse;
 import com.digginroom.digginroom.service.dto.RoomsResponse;
 import com.digginroom.digginroom.service.dto.TrackResponse;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -33,11 +30,9 @@ public class RoomService {
     @Transactional(readOnly = true)
     public RoomResponse recommend(final Long memberId) {
         Member member = memberRepository.getMemberById(memberId);
-        List<MemberGenre> memberGenres = memberGenreRepository.findByMemberId(memberId);
-        validateHasMemberGenres(memberGenres);
 
         try {
-            Room recommendedRoom = roomRecommender.recommend(memberGenres);
+            Room recommendedRoom = roomRecommender.recommend(memberId);
 
             return new RoomResponse(
                     recommendedRoom.getId(),
@@ -48,12 +43,6 @@ public class RoomService {
             );
         } catch (RecommendException e) {
             return this.recommend(memberId);
-        }
-    }
-
-    private void validateHasMemberGenres(final List<MemberGenre> memberGenres) {
-        if(memberGenres.isEmpty()){
-            throw new EmptyFavoriteException();
         }
     }
